Made separate assets manager page folder. Separated assets table into a component. Assets are now fetched by AssetService.
