+++ WebApp/indoor-localization/src/App.tsx
@@ -2,10 +2,9 @@ import { Route, Routes } from "react-router-dom";
 import "./App.css";
 import Header from "./components/Header";
 import Dashboard from "./components/dashboard/Dashboard";
-import Reports from "./components/Reports";
 import FacilitiesManager from "./components/facilities-manager/FacilitiesManager";
 import ReportsPage from "./components/reports-page/ReportsPage";
-import AssetsManager from "./components/AssetsManager";
+import AssetsManager from "./components/assets-manager/AssetsManager";
 
 function App() {
   return (
